feat: added Header.jsx file, moved template files to separate folder
