+++ src/views/Events.js
@@ -1,7 +1,8 @@
-import React,{Component} from "react";
+import React, { Component } from "react";
 
 // core components
-import DemoNavbar from "components/Navbars/DemoNavbar.js";
+import Header from "components/Header";
+import Navigation from "components/Navigation";
 import SimpleFooter from "components/Footers/SimpleFooter.js";
 
 // yess let's get those animations
@@ -60,9 +61,9 @@ SwiperCore.use([Navigation, Pagination]);
 class Events extends React.Component {
 
   componentDidMount() {
-      document.documentElement.scrollTop = 0;
-      document.scrollingElement.scrollTop = 0;
-      this.refs.main.scrollTop = 0;
+    document.documentElement.scrollTop = 0;
+    document.scrollingElement.scrollTop = 0;
+    this.refs.main.scrollTop = 0;
   }
 
   // componentDidMount() {
@@ -81,145 +82,115 @@ class Events extends React.Component {
   //   });
   // }
 
-render() {
-  return (
-    <>
-      <DemoNavbar />
-      <main ref="main">
-        <div className="position-relative">
-          {/* shape Hero */}
-          <section className="section section-lg section-shaped pb-100">
-            <div className="shape shape-style-1 shape-primary">
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-              </div>
-            
-            {/* SVG separator */}
-            <div className="separator separator-bottom separator-skew">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                preserveAspectRatio="none"
-                version="1.1"
-                viewBox="0 0 2560 100"
-                x="0"
-                y="0"
-              >
-                <polygon
-                  className="fill-white"
-                  points="2560 0 2560 100 0 100"
-                />
-              </svg>
-            </div>
-          </section>
-          {/* 1st Hero Variation */}
-        </div>
-        <section className="section section-lg">
-          <Row className="justify-content-center text-center ">
-            <Col lg="8">       
-            <h1 class="animate__animated animate__zoomIn animate__fast"><h2 className="display-1">EVENTS</h2></h1>
-            </Col>                
-          </Row>
-          
-          <Container className="py-lg-md d-flex">
-            <Row className="justify-content-center text-center">
+  render() {
+    return (
+      <>
+        <Navigation />
+        <main ref="main">
+          <Header />
+
+          <section className="section section-lg">
+            <Row className="justify-content-center text-center ">
               <Col lg="8">
-                <p className="lead text-muted">
-                  At Co-op Soc, our top aims are to socialise and meet new people. We facilitate this through a wide range of social events, including our highly anticipated annual camp and ball.
-                </p>
+                <h1 class="animate__animated animate__zoomIn animate__fast"><h2 className="display-1">EVENTS</h2></h1>
               </Col>
             </Row>
+
+            <Container className="py-lg-md d-flex">
+              <Row className="justify-content-center text-center">
+                <Col lg="8">
+                  <p className="lead text-muted">
+                    At Co-op Soc, our top aims are to socialise and meet new people. We facilitate this through a wide range of social events, including our highly anticipated annual camp and ball.
+                  </p>
+                </Col>
+              </Row>
             </Container>
             <br></br>
-         
-        </section>
 
-        <br></br>
-        <br></br>
-        <br></br>
-        <br></br>
-        <br></br>
-        <section className="section section-lg pt-lg-0 mt--200">
-          
-          <Container>
+          </section>
 
-          <div class="row">
-            <div class="col-lg-4 col-md-12 mb-4 mb-lg-0">
-              <img
-                src={require("assets/img/home/camp.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+          <br></br>
+          <br></br>
+          <br></br>
+          <br></br>
+          <br></br>
+          <section className="section section-lg pt-lg-0 mt--200">
 
-              <img
-                src={require("assets/img/home/gals.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+            <Container>
 
-              <img
-                src={require("assets/img/home/funRun.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
-            </div>
+              <div class="row">
+                <div class="col-lg-4 col-md-12 mb-4 mb-lg-0">
+                  <img
+                    src={require("assets/img/home/camp.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
 
-            <div class="col-lg-4 mb-4 mb-lg-0">
-              <img
-                src={require("assets/img/home/galsBall.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+                  <img
+                    src={require("assets/img/home/gals.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
 
-              <img
-                src={require("assets/img/home/campWhole.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+                  <img
+                    src={require("assets/img/home/funRun.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
+                </div>
 
-              <img
-                src={require("assets/img/home/ball2.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
-            </div>
+                <div class="col-lg-4 mb-4 mb-lg-0">
+                  <img
+                    src={require("assets/img/home/galsBall.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
 
-            <div class="col-lg-4 mb-4 mb-lg-0">
-              <img
-                src={require("assets/img/home/gals2.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+                  <img
+                    src={require("assets/img/home/campWhole.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
 
-              <img
-                src={require("assets/img/home/funRun2.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
+                  <img
+                    src={require("assets/img/home/ball2.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
+                </div>
 
-              <img
-                src={require("assets/img/home/ytb.jpg")}
-                class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
-                alt=""
-              />
-            </div>
-          </div>
-        </Container>
-        </section>
-        
+                <div class="col-lg-4 mb-4 mb-lg-0">
+                  <img
+                    src={require("assets/img/home/gals2.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
+
+                  <img
+                    src={require("assets/img/home/funRun2.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
+
+                  <img
+                    src={require("assets/img/home/ytb.jpg")}
+                    class="w-100 shadow-1-strong rounded mb-4 card-lift--hover"
+                    alt=""
+                  />
+                </div>
+              </div>
+            </Container>
+          </section>
 
 
-      </main>
-      <SimpleFooter />
-    </>
-  );
+
+        </main>
+        <SimpleFooter />
+      </>
+    );
   }
 
-  
+
 }
