+++ src/views/Index.js
@@ -1,7 +1,8 @@
 import React from "react";
 
 // core components
-import DemoNavbar from "components/Navbars/DemoNavbar.js";
+import Header from "components/Header";
+import Navigation from "components/Navigation";
 import SimpleFooter from "components/Footers/SimpleFooter.js";
 
 // nodejs library that concatenates classes
@@ -29,7 +30,7 @@ import {
 import { Link } from "react-router-dom";
 
 // import css
-import '../assets/css/my.css'; 
+import '../assets/css/my.css';
 
 // Effects
 import Typewriter from 'typewriter-effect';
@@ -45,59 +46,59 @@ import Swiper from "swiper";
 // import Swiper styles
 import 'swiper/swiper-bundle.css';
 // core version + navigation, pagination modules:
-import SwiperCore, { Navigation, Pagination } from 'swiper/core';
+import SwiperCore, { Navigation as SNavigation, Pagination } from 'swiper/core';
 
 // configure Swiper to use modules
-SwiperCore.use([Navigation, Pagination]);
+SwiperCore.use([SNavigation, Pagination]);
 
 
 class Index extends React.Component {
   componentDidMount() {
     document.documentElement.scrollTop = 0;
     document.scrollingElement.scrollTop = 0;
-    this.refs.main.scrollTop = 0;    
+    this.refs.main.scrollTop = 0;
   }
-  
+
   render() {
     return (
       <>
-        <DemoNavbar />
+        <Navigation />
         <main ref="main">
-          
+
           <div className="position-relative">
             {/* shape Hero */}
             <section className="section section-lg section-shaped pb-250">
               <div className="shape shape-style-1 shape-primary">
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
-                <span className="floating"/>
+                <span className="floating" />
+                <span className="floating" />
+                <span className="floating" />
+                <span className="floating" />
+                <span className="floating" />
+                <span className="floating" />
+                <span className="floating" />
 
               </div>
               <Container className="py-lg-md d-flex">
                 <div className="col px-0">
-                <br></br>
+                  <br></br>
                   <Row>
-                  <Col lg="8" className="typeWriter">
-                    <Typewriter
-                      options={{
-                        loop: false,
-                      }}
-                      onInit={(typewriter) => {
-                        typewriter.typeString('UNSW CO-OP SOCIETY')
-                          .start();
-                      }}
-                    />
-                    <p className="lead text-white" style={{fontWeight : 'normal'}}>
+                    <Col lg="8" className="typeWriter">
+                      <Typewriter
+                        options={{
+                          loop: false,
+                        }}
+                        onInit={(typewriter) => {
+                          typewriter.typeString('UNSW CO-OP SOCIETY')
+                            .start();
+                        }}
+                      />
+                      <p className="lead text-white" style={{ fontWeight: 'normal' }}>
                         A society for co-ops, by co-ops.
-                    </p>
-                  </Col>
+                      </p>
+                    </Col>
                   </Row>
                 </div>
-                
+
               </Container>
               {/* SVG separator */}
               <div className="separator separator-bottom separator-skew">
@@ -128,26 +129,26 @@ class Index extends React.Component {
                       <Card className="card-lift--hover shadow border-0">
                         <CardBody className="py-5">
                           <div className="icon icon-shape icon-shape-index1 rounded-circle mb-4">
-                          <i className="ni ni-calendar-grid-58" />
+                            <i className="ni ni-calendar-grid-58" />
                           </div>
                           <div className="justify-content-center text-center ">
-                          <h6 className="text-uppercase">
-                          Social Events
-                          </h6>
-                          <p className="description mt-3">
-                          At the core of the society, our aims are to socialise and meet new people. We facilitate this through a wide range of social events,
-                           including an annual camp and ball.
-                          </p>
-                          <br></br>
+                            <h6 className="text-uppercase">
+                              Social Events
+                            </h6>
+                            <p className="description mt-3">
+                              At the core of the society, our aims are to socialise and meet new people. We facilitate this through a wide range of social events,
+                              including an annual camp and ball.
+                            </p>
+                            <br></br>
                           </div>
                           <div className="justify-content-center text-center ">
-                          <Button
-                            className="btn-icon mb-3 mb-sm-0"
-                            color="index1"
-                            href="#/events"
-                          >
-                            Learn more
-                          </Button>
+                            <Button
+                              className="btn-icon mb-3 mb-sm-0"
+                              color="index1"
+                              href="#/events"
+                            >
+                              Learn more
+                            </Button>
                           </div>
                         </CardBody>
                       </Card>
@@ -156,26 +157,26 @@ class Index extends React.Component {
                       <Card className="card-lift--hover shadow border-0">
                         <CardBody className="py-5">
                           <div className="icon icon-shape icon-shape-index2 rounded-circle mb-4">
-                          <i class="fa fa-pencil"></i>
+                            <i class="fa fa-pencil"></i>
                           </div>
                           <div className="justify-content-center text-center ">
-                          <h6 className="text-uppercase">
-                            Publications
-                          </h6>
-                          <p className="description mt-3">
-                          Our goal is to create a platform in which scholars past and present can share their experiences, learn something new, and stay connected.
+                            <h6 className="text-uppercase">
+                              Publications
+                            </h6>
+                            <p className="description mt-3">
+                              Our goal is to create a platform in which scholars past and present can share their experiences, learn something new, and stay connected.
 
-                          </p>
-                          <br></br>
+                            </p>
+                            <br></br>
                           </div>
                           <div className="justify-content-center text-center ">
-                          <Button
-                            className="mt-4"
-                            color="index2"
-                            href="#/publications"
-                          >
-                            Learn more
-                          </Button>
+                            <Button
+                              className="mt-4"
+                              color="index2"
+                              href="#/publications"
+                            >
+                              Learn more
+                            </Button>
                           </div>
                         </CardBody>
                       </Card>
@@ -184,26 +185,26 @@ class Index extends React.Component {
                       <Card className="card-lift--hover shadow border-0">
                         <CardBody className="py-5">
                           <div className="icon icon-shape icon-shape-index3 rounded-circle mb-4">
-                          <i className="fa fa-heart" />
+                            <i className="fa fa-heart" />
                           </div>
                           <div className="justify-content-center text-center ">
-                          <h6 className="text-uppercase">
-                          Charity Events
-                          </h6>
-                          <p className="description mt-3">
-                          The charity portfolio is an integral way for Co-op scholars to give back to the community through events
-                          that raise awareness for a diverse range of charities and social issues.
-                          </p>
-                          <br></br>
+                            <h6 className="text-uppercase">
+                              Charity Events
+                            </h6>
+                            <p className="description mt-3">
+                              The charity portfolio is an integral way for Co-op scholars to give back to the community through events
+                              that raise awareness for a diverse range of charities and social issues.
+                            </p>
+                            <br></br>
                           </div>
                           <div className="justify-content-center text-center ">
-                          <Button
-                            className="btn-icon mb-3 mb-sm-0"
-                            color="index3"
-                            href="#/charity"
-                          >
-                            Learn more
-                          </Button>
+                            <Button
+                              className="btn-icon mb-3 mb-sm-0"
+                              color="index3"
+                              href="#/charity"
+                            >
+                              Learn more
+                            </Button>
                           </div>
                         </CardBody>
                       </Card>
@@ -218,14 +219,14 @@ class Index extends React.Component {
           <br></br>
           <br></br>
 
-          
+
           <section className="section section-lg pt-lg-0 mt--200">
-          
-          <Container>
 
-          
+            <Container>
 
-          {/* <hr></hr>
+
+
+              {/* <hr></hr>
           <section class="section section-lg">
           <Row>
           <Col>
@@ -269,10 +270,10 @@ class Index extends React.Component {
           </Col> 
           </Row>
         </section> */}
-          {/* <hr></hr> */}
-          {/* <iframe src="https://open.spotify.com/embed/show/25fOObrHq0t3hPrmf8B8Dx" width="100%" height="232" frameBorder="0" allowtransparency="true" allow="encrypted-media"></iframe> */}
-          
-          </Container>
+              {/* <hr></hr> */}
+              {/* <iframe src="https://open.spotify.com/embed/show/25fOObrHq0t3hPrmf8B8Dx" width="100%" height="232" frameBorder="0" allowtransparency="true" allow="encrypted-media"></iframe> */}
+
+            </Container>
           </section>
