+++ template/Landing.js
@@ -16,11 +16,6 @@
 
 */
 import React from "react";
-
-// core components
-import DemoNavbar from "components/Navbars/DemoNavbar.js";
-import CardsFooter from "components/Footers/CardsFooter.js";
-
 // nodejs library that concatenates classes
 import classnames from "classnames";
 
@@ -41,11 +36,15 @@ import {
   Col
 } from "reactstrap";
 
+// core components
+import Navigation from "components/Navigation";
+import CardsFooter from "components/Footers/CardsFooter.js";
+
 // index page sections
-import Download from "./IndexSections/Download.js";
+import Download from "./components/Download.js";
 
-
-class Index extends React.Component {
+class Landing extends React.Component {
+  state = {};
   componentDidMount() {
     document.documentElement.scrollTop = 0;
     document.scrollingElement.scrollTop = 0;
@@ -54,7 +53,7 @@ class Index extends React.Component {
   render() {
     return (
       <>
-        <DemoNavbar />
+        <Navigation />
         <main ref="main">
           <div className="position-relative">
             {/* shape Hero */}
@@ -686,6 +685,175 @@ class Index extends React.Component {
               </Row>
             </Container>
           </section>
+          <section className="section section-lg pt-0">
+            <Container>
+              <Card className="bg-gradient-warning shadow-lg border-0">
+                <div className="p-5">
+                  <Row className="align-items-center">
+                    <Col lg="8">
+                      <h3 className="text-white">
+                        We made website building easier for you.
+                      </h3>
+                      <p className="lead text-white mt-3">
+                        I will be the leader of a company that ends up being
+                        worth billions of dollars, because I got the answers. I
+                        understand culture.
+                      </p>
+                    </Col>
+                    <Col className="ml-lg-auto" lg="3">
+                      <Button
+                        block
+                        className="btn-white"
+                        color="default"
+                        href="https://www.creative-tim.com/product/argon-design-system-react?ref=adsr-landing-page"
+                        size="lg"
+                      >
+                        Download React
+                      </Button>
+                    </Col>
+                  </Row>
+                </div>
+              </Card>
+            </Container>
+          </section>
+          <section className="section section-lg bg-gradient-default">
+            <Container className="pt-lg pb-300">
+              <Row className="text-center justify-content-center">
+                <Col lg="10">
+                  <h2 className="display-3 text-white">Build something</h2>
+                  <p className="lead text-white">
+                    According to the National Oceanic and Atmospheric
+                    Administration, Ted, Scambos, NSIDClead scentist, puts the
+                    potentially record low maximum sea ice extent tihs year down
+                    to low ice.
+                  </p>
+                </Col>
+              </Row>
+              <Row className="row-grid mt-5">
+                <Col lg="4">
+                  <div className="icon icon-lg icon-shape bg-gradient-white shadow rounded-circle text-primary">
+                    <i className="ni ni-settings text-primary" />
+                  </div>
+                  <h5 className="text-white mt-3">Building tools</h5>
+                  <p className="text-white mt-3">
+                    Some quick example text to build on the card title and make
+                    up the bulk of the card's content.
+                  </p>
+                </Col>
+                <Col lg="4">
+                  <div className="icon icon-lg icon-shape bg-gradient-white shadow rounded-circle text-primary">
+                    <i className="ni ni-ruler-pencil text-primary" />
+                  </div>
+                  <h5 className="text-white mt-3">Grow your market</h5>
+                  <p className="text-white mt-3">
+                    Some quick example text to build on the card title and make
+                    up the bulk of the card's content.
+                  </p>
+                </Col>
+                <Col lg="4">
+                  <div className="icon icon-lg icon-shape bg-gradient-white shadow rounded-circle text-primary">
+                    <i className="ni ni-atom text-primary" />
+                  </div>
+                  <h5 className="text-white mt-3">Launch time</h5>
+                  <p className="text-white mt-3">
+                    Some quick example text to build on the card title and make
+                    up the bulk of the card's content.
+                  </p>
+                </Col>
+              </Row>
+            </Container>
+            {/* SVG separator */}
+            <div className="separator separator-bottom separator-skew zindex-100">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                preserveAspectRatio="none"
+                version="1.1"
+                viewBox="0 0 2560 100"
+                x="0"
+                y="0"
+              >
+                <polygon
+                  className="fill-white"
+                  points="2560 0 2560 100 0 100"
+                />
+              </svg>
+            </div>
+          </section>
+          <section className="section section-lg pt-lg-0 section-contact-us">
+            <Container>
+              <Row className="justify-content-center mt--300">
+                <Col lg="8">
+                  <Card className="bg-gradient-secondary shadow">
+                    <CardBody className="p-lg-5">
+                      <h4 className="mb-1">Want to work with us?</h4>
+                      <p className="mt-0">
+                        Your project is very important to us.
+                      </p>
+                      <FormGroup
+                        className={classnames("mt-5", {
+                          focused: this.state.nameFocused
+                        })}
+                      >
+                        <InputGroup className="input-group-alternative">
+                          <InputGroupAddon addonType="prepend">
+                            <InputGroupText>
+                              <i className="ni ni-user-run" />
+                            </InputGroupText>
+                          </InputGroupAddon>
+                          <Input
+                            placeholder="Your name"
+                            type="text"
+                            onFocus={e => this.setState({ nameFocused: true })}
+                            onBlur={e => this.setState({ nameFocused: false })}
+                          />
+                        </InputGroup>
+                      </FormGroup>
+                      <FormGroup
+                        className={classnames({
+                          focused: this.state.emailFocused
+                        })}
+                      >
+                        <InputGroup className="input-group-alternative">
+                          <InputGroupAddon addonType="prepend">
+                            <InputGroupText>
+                              <i className="ni ni-email-83" />
+                            </InputGroupText>
+                          </InputGroupAddon>
+                          <Input
+                            placeholder="Email address"
+                            type="email"
+                            onFocus={e => this.setState({ emailFocused: true })}
+                            onBlur={e => this.setState({ emailFocused: false })}
+                          />
+                        </InputGroup>
+                      </FormGroup>
+                      <FormGroup className="mb-4">
+                        <Input
+                          className="form-control-alternative"
+                          cols="80"
+                          name="name"
+                          placeholder="Type a message..."
+                          rows="4"
+                          type="textarea"
+                        />
+                      </FormGroup>
+                      <div>
+                        <Button
+                          block
+                          className="btn-round"
+                          color="default"
+                          size="lg"
+                          type="button"
+                        >
+                          Send Message
+                        </Button>
+                      </div>
+                    </CardBody>
+                  </Card>
+                </Col>
+              </Row>
+            </Container>
+          </section>
           <Download />
         </main>
         <CardsFooter />
@@ -694,4 +862,4 @@ class Index extends React.Component {
   }
 }
 
-export default Index;
+export default Landing;
