+++ template/Login.js
@@ -35,10 +35,10 @@ import {
 } from "reactstrap";
 
 // core components
-import DemoNavbar from "components/Navbars/DemoNavbar.js";
+import Navigation from "components/Navigation";
 import SimpleFooter from "components/Footers/SimpleFooter.js";
 
-class Register extends React.Component {
+class Login extends React.Component {
   componentDidMount() {
     document.documentElement.scrollTop = 0;
     document.scrollingElement.scrollTop = 0;
@@ -47,7 +47,7 @@ class Register extends React.Component {
   render() {
     return (
       <>
-        <DemoNavbar />
+        <Navigation />
         <main ref="main">
           <section className="section section-shaped section-lg">
             <div className="shape shape-style-1 bg-gradient-default">
@@ -66,11 +66,11 @@ class Register extends React.Component {
                   <Card className="bg-secondary shadow border-0">
                     <CardHeader className="bg-white pb-5">
                       <div className="text-muted text-center mb-3">
-                        <small>Sign up with</small>
+                        <small>Sign in with</small>
                       </div>
-                      <div className="text-center">
+                      <div className="btn-wrapper text-center">
                         <Button
-                          className="btn-neutral btn-icon mr-4"
+                          className="btn-neutral btn-icon"
                           color="default"
                           href="#pablo"
                           onClick={e => e.preventDefault()}
@@ -101,21 +101,11 @@ class Register extends React.Component {
                     </CardHeader>
                     <CardBody className="px-lg-5 py-lg-5">
                       <div className="text-center text-muted mb-4">
-                        <small>Or sign up with credentials</small>
+                        <small>Or sign in with credentials</small>
                       </div>
                       <Form role="form">
-                        <FormGroup>
-                          <InputGroup className="input-group-alternative mb-3">
-                            <InputGroupAddon addonType="prepend">
-                              <InputGroupText>
-                                <i className="ni ni-hat-3" />
-                              </InputGroupText>
-                            </InputGroupAddon>
-                            <Input placeholder="Name" type="text" />
-                          </InputGroup>
-                        </FormGroup>
-                        <FormGroup>
-                          <InputGroup className="input-group-alternative mb-3">
+                        <FormGroup className="mb-3">
+                          <InputGroup className="input-group-alternative">
                             <InputGroupAddon addonType="prepend">
                               <InputGroupText>
                                 <i className="ni ni-email-83" />
@@ -138,51 +128,51 @@ class Register extends React.Component {
                             />
                           </InputGroup>
                         </FormGroup>
-                        <div className="text-muted font-italic">
-                          <small>
-                            password strength:{" "}
-                            <span className="text-success font-weight-700">
-                              strong
-                            </span>
-                          </small>
+                        <div className="custom-control custom-control-alternative custom-checkbox">
+                          <input
+                            className="custom-control-input"
+                            id=" customCheckLogin"
+                            type="checkbox"
+                          />
+                          <label
+                            className="custom-control-label"
+                            htmlFor=" customCheckLogin"
+                          >
+                            <span>Remember me</span>
+                          </label>
                         </div>
-                        <Row className="my-4">
-                          <Col xs="12">
-                            <div className="custom-control custom-control-alternative custom-checkbox">
-                              <input
-                                className="custom-control-input"
-                                id="customCheckRegister"
-                                type="checkbox"
-                              />
-                              <label
-                                className="custom-control-label"
-                                htmlFor="customCheckRegister"
-                              >
-                                <span>
-                                  I agree with the{" "}
-                                  <a
-                                    href="#pablo"
-                                    onClick={e => e.preventDefault()}
-                                  >
-                                    Privacy Policy
-                                  </a>
-                                </span>
-                              </label>
-                            </div>
-                          </Col>
-                        </Row>
                         <div className="text-center">
                           <Button
-                            className="mt-4"
+                            className="my-4"
                             color="primary"
                             type="button"
                           >
-                            Create account
+                            Sign in
                           </Button>
                         </div>
                       </Form>
                     </CardBody>
                   </Card>
+                  <Row className="mt-3">
+                    <Col xs="6">
+                      <a
+                        className="text-light"
+                        href="#pablo"
+                        onClick={e => e.preventDefault()}
+                      >
+                        <small>Forgot password?</small>
+                      </a>
+                    </Col>
+                    <Col className="text-right" xs="6">
+                      <a
+                        className="text-light"
+                        href="#pablo"
+                        onClick={e => e.preventDefault()}
+                      >
+                        <small>Create new account</small>
+                      </a>
+                    </Col>
+                  </Row>
                 </Col>
               </Row>
             </Container>
@@ -194,4 +184,4 @@ class Register extends React.Component {
   }
 }
 
-export default Register;
+export default Login;
