+++ template/components/Hero.js
